refactor: extract generic components, restructure folders
