+++ apps/oeth/src/App.tsx
@@ -1,6 +1,7 @@
 import { Container, Stack } from '@mui/material';
-import { TrackingProvider, trackSentryError } from '@origin/oeth/shared';
+import { trackEvent, trackPage, trackSentryError } from '@origin/oeth/shared';
 import { ErrorBoundary, ErrorPage } from '@origin/shared/components';
+import { TrackingProvider } from '@origin/shared/providers';
 import { Outlet } from 'react-router-dom';
 
 import { Topnav } from './components/Topnav';
@@ -11,7 +12,12 @@ export const App = () => {
       ErrorComponent={<ErrorPage height={1} width={1} />}
       onError={trackSentryError}
     >
-      <TrackingProvider>
+      <TrackingProvider
+        onPageChange={trackPage}
+        onWalletConnect={(connect_address, connect_wallet) => {
+          trackEvent({ name: 'connect', connect_address, connect_wallet });
+        }}
+      >
         <Stack minWidth={370}>
           <Topnav />
           <Container
