+++ libs/oeth/shared/src/components/index.ts
@@ -1,3 +1,1 @@
 export * from './ApyHeader';
-export * from './AccountPopover';
-export * from './PriceTolerancePopover';
