+++ libs/oeth/shared/src/index.ts
@@ -1,4 +1,3 @@
 export * from './components';
 export * from './clients';
 export * from './generated/graphql';
-export * from './providers';
