--- libs/oeth/shared/src/providers/index.ts
@@ -1,1 +1,0 @@
-export * from './TrackingProvider';
