+++ libs/shared/providers/src/index.ts
@@ -1,4 +1,5 @@
 export * from './activities';
+export * from './analytics';
 export * from './chart';
 export * from './console';
 export * from './curve';
