+++ libs/shared/providers/src/slippage/hooks.ts
@@ -1,0 +1,7 @@
+import { useLocalStorageValue } from '@react-hookz/web';
+
+export const useSlippage = () => {
+  return useLocalStorageValue('@originprotocol/oeth-slippage', {
+    defaultValue: 0.001,
+  });
+};
