+++ libs/shared/providers/src/slippage/index.ts
@@ -1,7 +1,2 @@
-import { useLocalStorageValue } from '@react-hookz/web';
-
-export const useSlippage = () => {
-  return useLocalStorageValue('@originprotocol/oeth-slippage', {
-    defaultValue: 0.001,
-  });
-};
+export * from './components/PriceTolerancePopover';
+export * from './hooks';
